Geofence init and template added to generated map page
